+++ utils/js/VersionUtils.js
@@ -19,9 +19,16 @@ export function checkVersion(version, minVersion, maxVersion) {
  * @return boolean
  */
 export function checkMinVersion(version, minVersion) {
-    // Treat snapshot versions as the release version
-    version = version.replace('-SNAPSHOT', '.0');
-    return version && minVersion && semver.satisfies(version, '>=' + minVersion);
+    return (
+        version &&
+        minVersion &&
+        semver.satisfies(
+            normalizeVersion(version),
+            '>=',
+            normalizeVersion(minVersion),
+            {includePrerelease: true}
+        )
+    );
 }
 
 /**
@@ -29,7 +36,24 @@ export function checkMinVersion(version, minVersion) {
  * @return boolean
  */
 export function checkMaxVersion(version, maxVersion) {
-    // Treat snapshot versions as the release version
-    version = version.replace('-SNAPSHOT', '.0');
-    return version && maxVersion && semver.satisfies(version, '<=' + maxVersion);
+    return (
+        version &&
+        maxVersion &&
+        semver.satisfies(
+            normalizeVersion(version),
+            '<=',
+            normalizeVersion(maxVersion),
+            {includePrerelease: true}
+        )
+    );
+}
+
+/**
+ * Normalizes a Java/Maven style x.0-SNAPSHOT version to a semver compatible x.0.0-SNAPSHOT string.
+ * @param {string} version
+ * @return {string} - normalized version, if input matched as above, or input version unmodified.
+ */
+export function normalizeVersion(version) {
+    const isTwoDigitSnap = /^\d+\.0-SNAPSHOT$/.test(version);
+    return isTwoDigitSnap ? version.replace('-SNAPSHOT', '.0-SNAPSHOT') : version;
 }